HOTFIX Shortening /sportForecast
